Validation: move KeysIn to its own file
